Revert to previous state implementation
